requirements.txt
numpy==1.25.0
wandb==0.16.6
PyYAML==6.0.2
ray[tune]==2.9.3
prometheus_client==0.20.0

torch==2.1.1+cu121
torchvision==0.16.1+cu121
torchaudio==2.1.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121